requirements.txt
requests
openai
pypdf2
gradio
Python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
openai-agents